requirements.txt
Django<=1.12
django-cors-headers
djangorestframework
PyJWT
PyMySQL
pyotp
requests
